Bug/ con 2960

+++ cmd/csi-driver/conform/entrypoint.sh
@@ -4,22 +4,35 @@
 set -xe
 
 # Check if we are running as node-plugin
-for arg in "$@"
-do
+for arg in "$@"; do
     if [ "$arg" = "--node-service" ]; then
-        nodeservice=true
+        nodeService=true
     fi
 done
 
-if [ "$nodeservice" = true ]; then
-    if [ "${DISABLE_NODE_CONFORMANCE}" = "true" ]; then
-        echo "node conformance checks are disabled"
-    else
+# Set up node configurations based on environment variables
+disableNodeConformance=${DISABLE_NODE_CONFORMANCE}
+disableNodeConfiguraton=${DISABLE_NODE_CONFIGURATION}
+
+if [ "$nodeService" = true ]; then
+    # Disable the conformance checks
+    if [ "$disableNodeConformance" = "true" ]; then
+        echo "Node conformance checks are disabled"
+        disableConformanceCheck=true
+    fi
+
+    # Completely disable the node configuration by the driver.
+    if [ "$disableNodeConfiguraton" = "true" ]; then
+        echo "Node configuration is disabled"
+        disableConformanceCheck=true
+    fi
+
+    if [ "$disableConformanceCheck" != true ]; then
         # Copy HPE Storage Node Conformance checks and conf in place
         cp -f "/opt/hpe-storage/lib/hpe-storage-node.service" \
-          /etc/systemd/system/hpe-storage-node.service
+            /etc/systemd/system/hpe-storage-node.service
         cp -f "/opt/hpe-storage/lib/hpe-storage-node.sh" \
-          /etc/hpe-storage/hpe-storage-node.sh
+            /etc/hpe-storage/hpe-storage-node.sh
         chmod +x /etc/hpe-storage/hpe-storage-node.sh
 
         echo "running node conformance checks..."
@@ -32,22 +45,26 @@ if [ "$nodeservice" = true ]; then
     echo "copying hpe log collector diag script"
     cp -f "/opt/hpe-storage/bin/hpe-logcollector.sh" \
         /usr/local/bin/hpe-logcollector.sh
-    chmod +x  /usr/local/bin/hpe-logcollector.sh
+    chmod +x /usr/local/bin/hpe-logcollector.sh
 
     # Copy /etc/multipath.conf template if missing on host
-    if [ ! -f /host/etc/multipath.conf ]; then
+    if [ ! -f /host/etc/multipath.conf ] &&
+        [ "$disableNodeConfiguraton" != true ]; then
         cp /opt/hpe-storage/nimbletune/multipath.conf.upstream /host/etc/multipath.conf
     fi
+
     # symlink to host iscsi/multipath config files
     ln -s /host/etc/multipath.conf /etc/multipath.conf
     ln -s /host/etc/multipath /etc/multipath
     ln -s /host/etc/iscsi /etc/iscsi
+
     # symlink to host os release files for parsing
     if [ -f /host/etc/redhat-release ]; then
         # remove existing file from ubi
         rm /etc/redhat-release
         ln -s /host/etc/redhat-release /etc/redhat-release
     fi
+
     if [ -f /host/etc/os-release ]; then
         # remove existing file from ubi
         rm /etc/os-release
